-Updates to make Juggler compile better under Linux -Rewrote the reconfiguration system -Eliminated several compiler warnings
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@1504 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/modules/gadgeteer/gadget/Devices/Sim/SimAnalog.h
@@ -43,7 +43,7 @@ public:
    //: Return analog data
    virtual int getAnalogData(int devNum=0)
    {
-      vjASSERT(devNum < mAnaData.size());    // Make sure we have enough space
+      vjASSERT(devNum < (int)mAnaData.size());    // Make sure we have enough space
       return mAnaData[devNum];
    }
